requirements.txt
django-static-fontawesome
Django==4.0.3
djangorestframework~=3.13.1
Pillow~=9.0.1